Change 5
This is the main guts of the loading of the NPC's. What happens is now
Rooms also have a number that tells how many NPC's are there. If that
number is greater than zero it loads the NPC's from a file using the
room's name as the key to load the right enemies into the room.