Items and input fixes.
Add space and samples for new Items. Currently no space for them to be
placed.
With the previous commit, I realized capitalization was not being
handled correctly. Fixed all instances where the tolower function was
being used incorrectly. K is the same as k and KiLl is the same as kill
now. No need for capitalized variants of syntaxes.