Added Equipment and Examine commands.
Equipment displays worn equipment.
Changed inventory to be a Command rather than a PC function.
Examine allows the player to choose a target, inventory, equipment, or
look, and see all the names of the things that are seen there.
Ex: typing inventory displays
You are carrying 1 item.
a simple wooden sword
typing exa inv displays
You are carrying 1 item.
(woodsword) a simple wooden sword
woodsword is the name of the item the player has to type when using
commands such as 'wear' or 'drop', so this will become valuable later
on.