Created wear and remove commands.
Wear takes an item name from the inventory, removes it from the
inventory, then places it in the appropriate wear slot. Should remove a
worn item if it is in the way of the new item.
Remove command takes worn item name and moves it to the inventory.
Placed a wooden sword in the player's inventory for testing.