Some basic changes
I added a new ability and started to put the sound calls into the game.
In command functions I got rid of a line we don't need and in Player I
fixed an error that I found that we can't start the player with 0 AS as
it causes problems in the combat function so for now I just gave them 5
for a start.