Use command and no more healing after combat
Added a 'use' command and two variables for items if they are intended
to be 'use' items. Use items can only have the wear slot of "none".
In order to get use items working, I needed to fix the issue that caused
the player to heal after combat. This was due to all of the functions
surrounding the Character class' m_stats variable dealing with Combatant
return types. I went throughout all of the code and made them all return
Combatant* data types. By using pointers, the SetHp (and similar)
functions now set the HP of the Combatant instance being pointed to,
rather than a new copy created by the GetStats() function.