Added abilities to combat, plus bufixes
They currently ignore armor score and guard status.
Fixed the issue where they were only using the stats of the PC when each
ability was added.
Fixed an issue where when checking if a room was empty of NPCs, I was
calling an unrelated function, causing you to be trapped in a room after
killing an NPC.
Made guarding halve damage from the next attack.
TEMPORARY: Removed EXP from the NPC loading function (ruins the rest of
the stats if the NPC text file is not done properly, which at the
moment, it is not) to make combat testing possible.
TEMPORARY: Gave the PC 50 level-ups at the start for testing combat more
easily.