NPC Names and stats
Changed the NPC names and stats. Added to the help page so people will
know to type what is in the brackets to attack. Changed level to 1 to
test. Put Level Start comment in so we all know where the start level
manipulation spot is. Could not test past first guy do to bug. I believe
it is in the level up portion of code but not sure. Will edit bug
tracking when I can get it to open for me.