Tweaks to match new NPC existence tracking
Also got rid of the GetName call in look. There is an examine function
that serves this purpose. There is an examine function. It lets you
modify the look, equipment, and inventory commands. It puts the names in
front of things only when the player wants to see them. It was added a
week ago. "Added Equipment and Examine commands."